Initial commit
This is an export from a subversion repository at
http://svn.ashlock.us/public/georeport_ushahidi/   This is an autonomous
codeigniter app that is meant to sit along side Ushahidi and connect to
the same database. Edit your database settings in
/codeigniter/app/config/database.php   Copy the database.php.sample and
rename it database.php and edit your own settings. Do not check in
database.php into git  This was built to work with Ushahidi 2.0.1 and
Open311 GeoReport v2  Not all of the functionality of GeoReport v2 has
been developed here on this initial commit.